Annotations.
The use-case here is that you have a Xournal document, you'd
like to put textual comments in it, but you want it to be easy
to strip out those comments when exporting to PDF (say you're working
on a presentation).  So it would be handy to have a different type
of element which has exactly this toggleability.

In principle, we can apply this to all items, but text is what
I really needed.

As a visual cue that it's annotation, I apply a transparency. There
might be better ways to do it.  This transparency is NOT APPLIED on
PDF generation.

Button/cursor image choice is atrocious.

An alternate way to nirvana is to have a different layer mechanism
which is not per page but global.  This seemed a bit of fiddling
about.

File format extension: new 'annot' attribute on text elements.

Copy paste format modified.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>